refactor: move files to correct folder structure
