+++ packages/frontend/src/features/sync/components/SyncModalDelete.tsx
@@ -1,13 +1,13 @@
 import { NonIdealState, Spinner } from '@blueprintjs/core';
 import { Button, Group, Stack, Text } from '@mantine/core';
 import { useCallback, useEffect } from 'react';
-import ErrorState from '../../../../../components/common/ErrorState';
-import { useScheduler } from '../../../../../hooks/scheduler/useScheduler';
-import { useSchedulersDeleteMutation } from '../../../../../hooks/scheduler/useSchedulersDeleteMutation';
+import ErrorState from '../../../components/common/ErrorState';
+import { useScheduler } from '../../../hooks/scheduler/useScheduler';
+import { useSchedulersDeleteMutation } from '../../../hooks/scheduler/useSchedulersDeleteMutation';
 import {
     SyncWithGoogleSheetsModalAction,
     useSyncWithGoogleSheetsModal,
-} from './providers/SyncWithGoogleSheetsModalProvider';
+} from '../providers/SyncWithGoogleSheetsModalProvider';
 
 export const SyncModalDelete = () => {
     const { currentSchedulerUuid, setAction } = useSyncWithGoogleSheetsModal();
