+++ packages/frontend/src/features/sync/components/SyncModalForm.tsx
@@ -16,18 +16,18 @@ import {
 import { IconCirclesRelation } from '@tabler/icons-react';
 import { FC, useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import ErrorState from '../../../../../components/common/ErrorState';
-import MantineIcon from '../../../../../components/common/MantineIcon';
-import CronInput from '../../../../../components/ReactHookForm/CronInput';
-import { useChartSchedulerCreateMutation } from '../../../../../hooks/scheduler/useChartSchedulers';
-import { useScheduler } from '../../../../../hooks/scheduler/useScheduler';
-import { useSchedulersUpdateMutation } from '../../../../../hooks/scheduler/useSchedulersUpdateMutation';
-import { isInvalidCronExpression } from '../../../../../utils/fieldValidators';
-import { SelectGoogleSheetButton } from '../SelectGoogleSheetButton';
+import ErrorState from '../../../components/common/ErrorState';
+import MantineIcon from '../../../components/common/MantineIcon';
+import CronInput from '../../../components/ReactHookForm/CronInput';
+import { useChartSchedulerCreateMutation } from '../../../hooks/scheduler/useChartSchedulers';
+import { useScheduler } from '../../../hooks/scheduler/useScheduler';
+import { useSchedulersUpdateMutation } from '../../../hooks/scheduler/useSchedulersUpdateMutation';
+import { isInvalidCronExpression } from '../../../utils/fieldValidators';
 import {
     SyncWithGoogleSheetsModalAction,
     useSyncWithGoogleSheetsModal,
-} from './providers/SyncWithGoogleSheetsModalProvider';
+} from '../providers/SyncWithGoogleSheetsModalProvider';
+import { SelectGoogleSheetButton } from './SelectGoogleSheetButton';
 
 export const SyncModalForm: FC<{ chartUuid: string }> = ({ chartUuid }) => {
     const { action, setAction, currentSchedulerUuid } =
