+++ packages/frontend/src/features/sync/components/SyncModalView.tsx
@@ -12,12 +12,12 @@ import {
 import { IconDots, IconPencil, IconTrash } from '@tabler/icons-react';
 import cronstrue from 'cronstrue';
 import { FC } from 'react';
-import MantineIcon from '../../../../../components/common/MantineIcon';
-import { useChartSchedulers } from '../../../../../hooks/scheduler/useChartSchedulers';
+import MantineIcon from '../../../components/common/MantineIcon';
+import { useChartSchedulers } from '../../../hooks/scheduler/useChartSchedulers';
 import {
     SyncWithGoogleSheetsModalAction,
     useSyncWithGoogleSheetsModal,
-} from './providers/SyncWithGoogleSheetsModalProvider';
+} from '../providers/SyncWithGoogleSheetsModalProvider';
 
 export const SyncModalView: FC<{ chartUuid: string }> = ({ chartUuid }) => {
     const { data } = useChartSchedulers(chartUuid);
